backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
supabase==2.5.0
stripe==10.5.0
pydantic==2.8.2
gunicorn==22.0.0
pyjwt==2.10.1
bleach==6.2.0